api/requirements.txt
blinker==1.6.3
boto3==1.28.76
botocore==1.31.76
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
hiredis==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
urllib3==2.0.7
Werkzeug==3.0.1
